some quick testing and some notes for myself.
